requirements.txt
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
aylien-apiclient==0.7.0
aylien-news-api==5.2.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
Flask==3.1.0
Flask-Caching==2.3.1
flask-caching>=2.0.0
flask-cors==5.0.1
fsspec==2025.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
openai==1.64.0
packaging==24.2
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytrends>=4.9.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
python-dotenv==1.0.0
pytrends
aiohttp==3.9.3
quart==0.19.4
quart-cors==0.7.0
hypercorn==0.16.0
